requirements.txt
absl-py==0.10.0
alabaster==0.7.12
appdirs==1.4.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
astor==0.8.1
astroid==2.9.3
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
awscli==1.18.56
Babel==2.9.1
backcall==0.2.0
backports.weakref==1.0rc1
bcrypt==3.1.7
beautifulsoup4==4.10.0
bleach==1.5.0
blinker==1.4
bokeh==2.3.3
boto==2.49.0
boto3==1.21.21
botocore==1.16.6
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
cloud-init==21.4
cmake==3.22.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
coremltools==3.1
cpplint==1.6.0
cryptography==2.8
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
docutils==0.15.2
dparse==0.5.1
ec2-hibinit-agent==1.0.0
entrypoints==0.4
environment-kernels==1.1.1
filelock==3.0.4
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gdown==4.4.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.19.1
grpcio==1.32.0
h5py==2.10.0
hibagent==1.0.1
html5lib==0.9999999
httplib2==0.20.4
huggingface-hub==0.4.0
hyperlink==21.0.0
idna==3.3
imageio==2.15.0
imagesize==1.3.0
importlib-metadata==3.10.1
importlib-resources==5.4.0
incremental==21.3.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.17.2
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
language-selector==0.1
lazy-object-proxy==1.7.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mock==4.0.3
more-itertools==8.12.0
mpmath==1.2.1
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.10.4
networkx==2.5.1
nose==1.3.7
notebook==6.4.10
numba==0.53.1
numpy==1.19.5
nvidia-ml-py==11.450.51
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
PAM==0.4.2
pandas==0.25.3
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
platformdirs==2.4.0
prettytable==0.7.2
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.11.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycrypto==2.6.1
pydot==1.4.2
pygal==3.0.0
Pygments==2.11.2
PyGObject==3.26.1
PyJWT==1.5.3
pylint==2.12.2
PyNaCl==1.3.0
pyOpenSSL==17.5.0
pyparsing==3.0.7
pyrsistent==0.18.0
pyserial==3.4
PySocks==1.7.1
pystache==0.6.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
pytorch-lightning==0.8.5
pytz==2021.3
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2021.8.3
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.1.5
resampy==0.2.1
rsa==3.4.2
s3transfer==0.3.7
sacremoses==0.0.43
safety==1.10.3
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.11.2
Send2Trash==1.8.0
service-identity==16.0.0
six==1.15.0
snowballstemmer==2.2.0
sos==4.3
soupsieve==2.3.1
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssh-import-id==5.7
sympy==1.9
systemd-python==234
tb-nightly==2.3.0a20200415
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow-io==0.17.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
tf-estimator-nightly==2.3.0.dev2020041501
threadpoolctl==3.1.0
tifffile==2020.9.3
tokenizers==0.10.3
toml==0.10.2
torchvision==0.5.0
tornado==6.1
tqdm==4.63.0
traitlets==4.3.3
transformers==4.11.3
turicreate==6.0
Twisted==22.2.0
typed-ast==1.5.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.6.0
wrapt==1.12.1
zipp==3.6.0
zope.interface==5.4.0
